refactor: reorganize into identity and integrity modules (closes #52)
Milestone 1.1 — split the flat crypto/chain modules into two clearly
scoped public modules within crates/edgesentry-rs:

- src/identity.rs  — Ed25519 signing and verification (sign_payload_hash,
  verify_payload_signature); previously in src/crypto.rs
- src/integrity.rs — BLAKE3 payload hashing and hash-chain verification
  (compute_payload_hash, verify_chain, ChainError); previously split
  across src/crypto.rs and src/chain.rs

Internal callers (agent.rs, ingest/storage.rs, ingest/verify.rs) updated
to import from the new modules. The public API of edgesentry_rs is
unchanged: all symbols remain re-exported from lib.rs. Both modules are
now pub, making edgesentry_rs::identity and edgesentry_rs::integrity
directly accessible to library consumers.

Also updates roadmap.md to reflect M1.1 as complete.
