+++ crates/edgesentry-rs/src/agent.rs
@@ -1,6 +1,7 @@
 use ed25519_dalek::SigningKey;
 
-use crate::crypto::{compute_payload_hash, sign_payload_hash};
+use crate::identity::sign_payload_hash;
+use crate::integrity::compute_payload_hash;
 use crate::record::{AuditRecord, Hash32};
 
 pub fn build_signed_record(
