+++ crates/edgesentry-rs/src/integrity.rs
@@ -1,7 +1,19 @@
+//! BLAKE3 hash-chain integrity — payload hashing and chain verification.
+//!
+//! This module covers the tamper-detection layer: hashing a raw payload with
+//! BLAKE3 and verifying that a sequence of `AuditRecord`s forms an unbroken
+//! hash chain.
+
 use thiserror::Error;
 
 use crate::record::AuditRecord;
 
+/// Compute the BLAKE3 hash of a raw payload.
+pub fn compute_payload_hash(payload: &[u8]) -> [u8; 32] {
+    *blake3::hash(payload).as_bytes()
+}
+
+/// Errors produced by [`verify_chain`].
 #[derive(Debug, Error, PartialEq, Eq)]
 pub enum ChainError {
     #[error("invalid previous hash at index {index}")]
@@ -14,6 +26,12 @@ pub enum ChainError {
     },
 }
 
+/// Verify that `records` form a valid hash chain.
+///
+/// - The first record must have `prev_record_hash == [0u8; 32]`.
+/// - Each subsequent record's `prev_record_hash` must equal the hash of the
+///   preceding record.
+/// - Sequences must be strictly monotonically increasing by 1.
 pub fn verify_chain(records: &[AuditRecord]) -> Result<(), ChainError> {
     if records.is_empty() {
         return Ok(());
